backend/requirements.txt
Flask==2.2.2
beautifulsoup4==4.11.1
urllib3==1.26.14
requests==2.28.2
gspread==5.7.2
gspread-formatting==1.1.2
python-dotenv==1.0.0
Flask-Cors==4.0.0
